Retrieve event by its ID
Sometimes our Eventbrite account isn't the sole source of the registration
pages. For instance, we are a subuser under General Assembly's EB account.
Since subusers are not allowed to set up their own API keys, the easiest way to
access an event is to just retrieve the event by its ID (which can be found in
the EB url). This eliminates the need to reach through an organizer and their
events.