Introduce an event finder
Now that we're just retrieving a single event, there's no real need for there to
be an `Eventbrite` class on its own. Instead, it's been transformed into a
EventFinder object. Eventually, we'll also want to create a similar finder
object for venues, so we've abstracted away common logic into `Finder`, which it
can then inherit from.