Simplify how we specify dependencies in our classes and their specs
Each class and their correspending unit specs should really only have to specify
dependencies that are directly related to them without having to reach too far.
For example, the `EventFinder` spec should only have to require the file
containing its class and nothing else that the class itself actually depends on,
like `Eventbrite::Event`.

On the other hand, we could also move some spec config over to the
`spec_helper.rb` file and have `rails_helper` refer to it, so that that can be
centralized and shared across all tests.